refactor: Create save agent directory
