+++ app/projects/new/action.ts
@@ -2,25 +2,15 @@
 
 import pgvector from 'pgvector'
 import { v4 as uuidv4 } from 'uuid'
+import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
+import { openai } from '@ai-sdk/openai'
 import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3'
 import { fromCognitoIdentityPool } from '@aws-sdk/credential-providers'
 import { PrismaClient } from '@prisma/client'
-import { auth } from "../../auth"
-import { revalidatePath } from 'next/cache'
-
-import { openai } from '@ai-sdk/openai'
-import { mkdtempSync, mkdirSync, rmSync, writeFileSync } from 'fs'
-import { tmpdir } from 'os'
-import { MultiFileLoader } from 'langchain/document_loaders/fs/multi_file'
-import { dirname, extname, join } from 'path'
-import { CSVLoader } from '@langchain/community/document_loaders/fs/csv'
-import { DocxLoader } from '@langchain/community/document_loaders/fs/docx'
-import { JSONLoader } from 'langchain/document_loaders/fs/json'
-import { PDFLoader } from '@langchain/community/document_loaders/fs/pdf'
-import { PPTXLoader } from '@langchain/community/document_loaders/fs/pptx'
-import { TextLoader } from 'langchain/document_loaders/fs/text'
-import { generateFilesObjectAgent, streamSummaryTextAgent } from './agents'
+import { generateFilesObjectAgent, streamSummaryTextAgent } from './_agent/agents'
+import { filesLoader } from './_agent/loader'
+import { auth } from '../../auth'
 
 const prisma = new PrismaClient()
 const s3Client = new S3Client({
@@ -38,60 +28,20 @@ export const confirmAction = async (
 ) => {
   // 重要なファイルをピックアップ
   const paths = await generateFilesObjectAgent(webkitRelativePaths)
-
-  // フォームデータを取得
   const formFiles = formData.getAll('files') as File[]
   const files = formFiles
-    // formFileにwebkitRelativePathが含まれないため手動で追加
     .map((file, i) => {
+      // formFileにwebkitRelativePathが含まれないため手動で追加
       const webkitRelativePath = webkitRelativePaths[i]
       return Object.assign(file, { webkitRelativePath })  // readonlyを上書き
     })
-    // 重要なファイルのformFileのみを取得
-    .filter((file) => paths.includes(file.webkitRelativePath))
+    .filter((file) => {
+      // 重要なファイルのformFileのみを取得
+      return paths.includes(file.webkitRelativePath)
+    })
 
-  // 一時ファイルを作成
-  const prefix = join(tmpdir(), 'savepoint-')
-  const tmpDir = mkdtempSync(prefix)
-  for (const file of files) {
-    const filename = file.name
-    const dirName = dirname(file.webkitRelativePath)
-    const buffer = Buffer.from(await file.arrayBuffer())
-
-    const fileDir = join(tmpDir, dirName)
-    const filePath = join(fileDir, filename)
-    mkdirSync(fileDir, { recursive: true })
-    writeFileSync(filePath, buffer)
-  }
-
-  // ドキュメントを読み込み
-  const loaders = {
-    '.csv': (path: string) => new CSVLoader(path),
-    '.docx': (path: string) => new DocxLoader(path),
-    '.json': (path: string) => new JSONLoader(path),
-    '.pdf': (path: string) => new PDFLoader(path),
-    '.pptx': (path: string) => new PPTXLoader(path),
-    '.txt': (path: string) => new TextLoader(path),
-  }
-
-  const loader = new MultiFileLoader(
-    paths
-      // loadersに拡張子が定義されているファイルのみを抽出
-      .filter((path: string) => extname(path) in loaders)
-      .map((path) => join(tmpDir, path)),
-    loaders,
-  )
-
-  const documents = await loader.load()
-  const content = documents.reduce(
-    (content, document) => content + '\n\n' + document.pageContent,
-    ''  // 初期値
-  )
-
-  // 一時ファイルを削除
-  rmSync(tmpDir, { recursive: true })
-  
-  // プロジェクト概要を生成
+  // ファイルからプロジェクト概要を生成
+  const content = await filesLoader(files)
   const summaryStream = await streamSummaryTextAgent(content)
   return summaryStream
 }
